Move Data-Driven MPC control loop to dedicated module
- Move the `simulate_data_driven_mpc_control_loop` function from
  `controller_operation.py` to the new module
  `data_driven_mpc_sim.py`.
- Update import statements in example files to reflect this
  change.

This change improves the project structure by isolating the core
Data-Driven MPC control loop logic into a dedicated file.
