+++ examples/lti_control/lti_dd_mpc_example.py
@@ -38,7 +38,8 @@ from utilities.controller.controller_params import (
 from utilities.controller.controller_creation import (
     create_lti_data_driven_mpc_controller)
 from utilities.controller.controller_operation import (
-    randomize_initial_system_state, generate_initial_input_output_data,
+    randomize_initial_system_state, generate_initial_input_output_data)
+from utilities.controller.data_driven_mpc_sim import (
     simulate_data_driven_mpc_control_loop)
 
 from utilities.visualization.data_visualization import (
