+++ utilities/reproduction/paper_reproduction.py
@@ -9,7 +9,7 @@ from utilities.controller.controller_params import (
     LTIDataDrivenMPCParamsDictType)
 from utilities.controller.controller_creation import (
     create_lti_data_driven_mpc_controller)
-from utilities.controller.controller_operation import (
+from utilities.controller.data_driven_mpc_sim import (
     simulate_data_driven_mpc_control_loop)
 from utilities.visualization.data_visualization import (
     plot_input_output, create_input_output_figure)
